% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_model.R
\name{compute_transition_matrix.default}
\alias{compute_transition_matrix.default}
\title{Compute a single transition matrix
The first matrix index is 'from', second is 'to'}
\usage{
\method{compute_transition_matrix}{default}(trans_model, predictor_values)
}
\description{
Compute a single transition matrix
The first matrix index is 'from', second is 'to'
}
